Created first service

+++ fanbase-data/src/main/java/pl/edu/pjwstk/fanbasedata/repositories/FilmRepository.java
@@ -3,5 +3,5 @@ package pl.edu.pjwstk.fanbasedata.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.edu.pjwstk.fanbasedata.model.Film;
 
-public interface FilmRepository extends JpaRepository<Film, Integer> {
+public interface FilmRepository extends JpaRepository<Film, Long> {
 }
