+++ fanbase-data/src/main/java/pl/edu/pjwstk/fanbasedata/repositories/SpeciesRepository.java
@@ -3,5 +3,5 @@ package pl.edu.pjwstk.fanbasedata.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.edu.pjwstk.fanbasedata.model.Species;
 
-public interface SpeciesRepository extends JpaRepository<Species, Integer> {
+public interface SpeciesRepository extends JpaRepository<Species, Long> {
 }
